// web/src/__generated__/schemas/ZoneView.ts
/**
 * ⚠️ THIS FILE IS AUTO-GENERATED. DO NOT EDIT MANUALLY.
 * All manual edits will be lost when this file is regenerated.
 */

import type { ZoneOrigin } from './enums/ZoneOrigin';

export type ZoneView = { 
  displayName: string; 
  origin: ZoneOrigin; 
  plan: string; 
  zoneId: string; 
  archived?: boolean; 
  defaultRole?: string | null; 
  secret?: string | null; 
};